+++ lib/fuzz_cmd/src/run.ml
@@ -11,15 +11,13 @@
 
 open Core (* for Filename.arg_type *)
 
-open Act_common
-
 let write_trace (trace : Act_fuzz.Trace.t) (oc : Stdio.Out_channel.t) :
     unit Or_error.t =
   Sexp.output_hum oc [%sexp (trace : Act_fuzz.Trace.t)] ;
   Result.ok_unit
 
 let run ?(seed : int option) ?(trace_output : string option)
-    (args : _ Toplevel.Args.With_files.t) (o : Output.t)
+    (args : _ Toplevel.Args.With_files.t) (o : Act_common.Output.t)
     (act_config : Act_config.Act.t) : unit Or_error.t =
   let config =
     act_config |> Act_config.Act.fuzz